backend/requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.13.0
Django==6.0.1
django-cors-headers==4.9.0
django-unfold==0.77.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
frozendict==2.4.7
idna==3.11
multitasking==0.0.12
numpy==2.4.2
pandas==3.0.0
peewee==3.19.0
platformdirs==4.5.1
protobuf==6.33.5
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
websockets>=15.0,<17.0
pyotp==2.9.0
qrcode[pil]==8.0
yfinance==1.1.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
supabase==2.13.0
